# docs/api.yml
openapi: 3.0.0
info:
  title: 'WeDigital Garden - Core API'
  description: 'Documentation of Core API for We Digital Garden'
  version: 0.0.1
servers:
  -
    url: 'http://localhost'
paths:
...
  /api/contact:
    post:
      tags:
        - Contact
      summary: 'Send a email with data of contact form'
      operationId: 'App\Http\Controllers\Api\ContactController::index'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
      responses:
        '200':
          description: 'Email is send'
  /api/contact/startup:
    post:
      tags:
        - Contact
      summary: 'Send a email with data of contact form'
      operationId: 'App\Http\Controllers\Api\ContactController::startup'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
      responses:
        '200':
          description: 'Email is send'
...
components:
...
  securitySchemes:
    bearer:
      type: http
      name: bearerAuth
      bearerFormat: JWT
      scheme: bearer